+++ node/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add config parameter `reader_max_connections` & make archive node options configurable [#3917]
+- Add support for range activation features
 
 ## [1.4.0] - 2025-11-06